space issue resolved in notes
