--- src/components/pages/notes/note.ts
@@ -1,10 +1,0 @@
-export interface Note {
-    id: number;
-    title: string;
-    content: string;
-    color: string;
-    is_pinned: boolean;
-    is_archived: boolean;
-    created_at: string;
-    updated_at: string;
-  }
